Modify unit test
