+++ src/test/java/unit/drivecommand/machine/MachineTest.java
@@ -1,4 +1,4 @@
-package test.drivecommand.machine;
+package unit.drivecommand.machine;
 
 import java.io.IOException;
