+++ src/test/java/unit/drivecommand/machine/input/RangefinderTest.java
@@ -1,4 +1,4 @@
-package test.drivecommand.machine.input;
+package unit.drivecommand.machine.input;
 
 import static org.testng.Assert.*;
 
@@ -11,30 +11,30 @@ import org.testng.annotations.Test;
 
 import com.pileproject.drivecommand.command.CommandBase;
 import com.pileproject.drivecommand.machine.DeviceType;
-import com.pileproject.drivecommand.machine.input.SoundSensor;
+import com.pileproject.drivecommand.machine.input.Rangefinder;
 import com.pileproject.drivecommand.model.ProtocolBase;
 
 @SuppressWarnings("serial")
-public class SoundSensorTest {
+public class RangefinderTest {
 	@Mocked private ProtocolBase protocol;
 	private final int PORT = 0;
-	private final int VALUE_DB = 100;
+	private final int VALUE_DISTANCE = 200;
 	private final String KEY_VALUE = "value";
 	
 	@Test
-	public void getSoundDb() {
+	public void getDistance() {
 		new Expectations() {{
 			protocol.exec(PORT, (CommandBase)any); 
-			result = new HashMap<String, Object>() {{put(KEY_VALUE, VALUE_DB);}};
+			result = new HashMap<String, Object>() {{put(KEY_VALUE, VALUE_DISTANCE);}};
 		}};
-		SoundSensor ss = new SoundSensor(PORT, protocol);
-		assertEquals(ss.getDb(), VALUE_DB);
+		Rangefinder rf = new Rangefinder(PORT, protocol);
+		assertEquals(rf.getDistance(), VALUE_DISTANCE);
 	}
 	
 	@Test
-	public void deviceTypeIsSoundSensor() {
-		SoundSensor ss = new SoundSensor(PORT, protocol);
-		assertEquals(ss.getDeviceType(), DeviceType.SOUND_SENSOR);
+	public void deviceTypeIsRangefinder() {
+		Rangefinder rf = new Rangefinder(PORT, protocol);
+		assertEquals(rf.getDeviceType(), DeviceType.RANGEFINDER);
 	}
 }
