--- src/test/java/unit/test/drivecommand/model/ev3/Ev3ProtocolTest.java
@@ -1,5 +1,0 @@
-package test.drivecommand.model.ev3;
-
-public class Ev3ProtocolTest {
-	
-}
